Fix Bug and Add Word Wrap
This commit fixes a bug that occurs when the user clicks a number, an operator, and then enter. It used to freeze so that nothing could be done after. Now, it prevents you from clicking enter until a second number is placed. Word wrap was also added to allow any amount of numbers to fit on the screen.